requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==7.0.1
Django==4.2.18
django-allauth==65.3.1
django-appconf==1.1.0
django-celery==3.1.17
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-compressor==4.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-imagekit==5.0.0
django-js-asset==2.2.0
django-qrcode==0.3
django-simple-history==3.8.0
django-timezone-field==7.1
django-tinymce==4.1.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
exceptiongroup==1.2.2
Faker==35.2.0
fonttools==4.56.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
incremental==24.7.2
kiwisolver==1.4.7
kombu==5.4.2
Markdown==3.7
matplotlib==3.9.4
msgpack==1.1.0
numpy==2.0.2
oauthlib==3.2.2
onesignal-sdk==2.0.0
opencv-python==4.11.0.86
packaging==24.2
pilkit==3.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
qrcode==8.0
rcssmin==1.1.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zipp==3.21.0
zope.interface==7.2
